copy codahale files to org.apache.bookkeeper.stats.codahale
